Fix import linting rule for moved util package
Update .golangci.yaml to reflect that util package is now at
internal/shared/util (not internal/util), and update all imports
to comply with the linting rule.
