+++ internal/catalogd/features/features.go
@@ -5,7 +5,7 @@ import (
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/component-base/featuregate"
 
-	fgutil "github.com/operator-framework/operator-controller/internal/shared/util/featuregates"
+	featuregatesutil "github.com/operator-framework/operator-controller/internal/shared/util/featuregates"
 )
 
 const (
@@ -24,5 +24,5 @@ func init() {
 
 // LogFeatureGateStates logs the state of all known feature gates for catalogd
 func LogFeatureGateStates(log logr.Logger, fg featuregate.FeatureGate) {
-	fgutil.LogFeatureGateStates(log, "catalogd feature gate status", fg, catalogdFeatureGates)
+	featuregatesutil.LogFeatureGateStates(log, "catalogd feature gate status", fg, catalogdFeatureGates)
 }
