+++ internal/shared/util/test/artifacts.go
@@ -1,4 +1,4 @@
-package testutils
+package test
 
 import (
 	"context"
