+++ internal/shared/util/test/summary.go
@@ -1,4 +1,4 @@
-package testutils
+package test
 
 import (
 	"context"
