+++ internal/shared/util/test/templates/mermaid_chart.md.tmpl
