+++ test/e2e/network_policy_test.go
@@ -15,7 +15,7 @@ import (
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	utils "github.com/operator-framework/operator-controller/internal/shared/util/testutils"
+	testutil "github.com/operator-framework/operator-controller/internal/shared/util/test"
 )
 
 const (
@@ -160,7 +160,7 @@ func TestNetworkPolicyJustifications(t *testing.T) {
 		}
 	}
 
-	clientForComponent := utils.FindK8sClient(t)
+	clientForComponent := testutil.FindK8sClient(t)
 
 	operatorControllerNamespace := getComponentNamespace(t, clientForComponent, operatorManagerSelector)
 	catalogDNamespace := getComponentNamespace(t, clientForComponent, catalogdManagerSelector)
